feat: add dataset loading
Add tag management---deletion & addition of tags with dynamically-generated buttons. Add Queue and Trie data structures for the program. Add dataset loading features. Users may load a directory as a dataset with the existing 'Load' button. The program will search recursively for all '*.png' images within the directory and, for each that has no corresponding caption (a '*.TXT' file with an identical name), users will be prompted in queue on whether or not they would like to create one (selecting 'No' means that the *.png file will be ignored in the load).

+++ tags.py
@@ -1,34 +1,44 @@
-from tkinter import Button, LEFT, END
+from tkinter import Button, LEFT, END, DISABLED
 
-class TagBox():
+class TagBox:
     def __init__(self, win, parent, tag):
         self.win = win
         self.parent = parent
         self.tag_text = tag
+        self.is_trigger = False
+        if self.win is not None:
+            if self.win.lora_in_training is not None:    
+                if tag == self.win.lora_in_training.trigger_word:
+                    self.is_trigger = True
         print(f'Generating button for tag: {tag}')
         self.bt = Button(parent, text=tag, command=self.devise_action)
-        #self.bt.pack(side=LEFT)
+        if self.is_trigger:
+            self.bt.config(text=f'{tag}🔒', bg='gold', state=DISABLED)
 
     def devise_action(self):
+        if self.is_trigger is True:
+            return
         match self.win.tag_click_mode.get():
             case "Delete":
                 self.delete()
+            case "Delete_All":
+                print(f'Removing tag "{self.tag_text}" from all .txt files in dataset {self.win.directory}')
+                self.win.lora_in_training.remove_tag_from_image_caption(self.tag_text, all=True)
             case "Apply_All":
-                self.proliferate_tag()
+                print(f'Applying tag "{self.tag_text}" to all .txt files in dataset {self.win.directory}')
+                self.win.lora_in_training.add_tag_to_image_caption(self.tag_text, all=True)
             case _:
                 raise ValueError("only 'Delete' and 'Apply_All' are acceptable actions")
-                print(self.win.tag_click_mode.get())
-                return
+        self.win.refresh()
     
     def delete(self):
         text = self.bt.cget("text")
         print(f'deleting tag in caption: {self.tag_text}')
-        tags = self.win._Window__caption_txt_field.get("1.0", "end-1c")
-        replaced = tags.replace(f"{self.tag_text}, ", "")
-        self.win._Window__caption_txt_field.delete("1.0", "end")
-        self.win._Window__caption_txt_field.insert(END, replaced)
-        self.bt.destroy()
+        tags = self.win._TrainLoraWin__caption_txt_field.get("1.0", "end-1c")
+        png_path = self.win.lora_in_training.image_set[self.win.image_set_display_index]
+        self.win.lora_in_training.remove_tag_from_image_caption(self.tag_text, png_path=png_path)
+        self.win.refresh()
+        self.destroy()
     
-    def proliferate_tag(self):
-        print(f'Applying tag "{self.tag_text}" to all .txt files in dataset {self.win.directory}')
-        return
+    def destroy(self):
+        self.bt.destroy()
